[SMA-1] Add support for scanning mutiple archives at a time
The inputPath may now point to a single archive or to a directory. In the event
of it pointing to a directory the directory structure is scanned and every
archive that's found is analyzed. An archive is identified by the file being in
Zip format, i.e. there's no requirement for an archive to have a well-known
file suffix for it to be analyzed.

A separate report is produced for each archive that's analyzed. Each report is
written into its own directory beneath the configured outputPath location. In
the case of multiple archives being scanned, the directory structure of the
reports that are produced will mirror the structure of the directory that was
provided as input.