[SMA-13] Handle input path and output path clashes
Depending on where the archive that is being analyzed exists on the filesystem
and the output path that the reports are being written to, it's possible for
the output to clash with the input. This commit improves the handling of
such an error case, and also improve the help information to explain how the
output path is used and where the reports are written to.