Move version declarations into root build.gradle
The Artifactory-based release process automatically updates the version that's
declared in gradle.properties. Unfortunately in doing so, it also removes
formatting from the file and  reorders the properties. To work  around this, we
now only declare the project's version in gradle.properties. The versions of
the project's dependencies are now declared in build.gradle where they will be
unaffected by the release process.