[SMA-22] Fix setting of CLASSPATH on Windows
The migration-analysis.bat script was setting the CLASSPATH by
concantenation but did not consider the initial case where the
CLASSPATH was empty. This resulted in CLASSPATH containing an empty
entry which is interpretted as the current working directory.
The knock-on effect is that StandardClassPathScanner will
recursively scan the current working directory. If the current
working directory happens to be inside the SMA installation this
is benign, however if the current working directory is outside the
SMA installation it leads to, at best, unnecessary scanning
during launch.

This commit addresses the issue by updating migration-analysis.bat
so that when CLASSPATH is empty it directly assigns it rather than
using concatenation, thereby avoiding the empty entry.