[SMA-23] Improve output directory naming
Prior to this commit, each report was written into a directory with
the same name as its input archive. This meant that, if the output
path was the same directory as the one that contained the input archive,
their names would clash.

To overcome this problem, each report is now written into a directory
with a name of the form migration-analysis-<input-archive>. For example,
if my-app.ear is analysed, its report will be written into a directory
named migration-analysis-my-app.ear.

In addition to the change described above, each report directory
now contains a subdirectory for the report type. Continuing the example
above, if an HTML report is produced it will be found in a directory
named migration-analysis-my-app.ear/html. This paves the way for
rendering multiple reports from the same analysis (SMA-3). For example,
it's currently expected that analysing my-app.ear with reports in
format1 and format2 would produce the following:

migration-analysis-my-app.ear/
    format1/
		…
	format2/
		…