Configure the source and target compatibility
Rather than relying on the project being built with the correct
JDK, the source and target compatibility is now explicitly configured
in the build. This should all the project to be built with Java 7
(which is required by the FindBugs plugin used on sonar.spring.io),
while still producing Java 6-compatible byte code.

See gh-9